+++ src/core_modules/capture-core/components/DataEntries/EnrollmentRegistrationEntry/EnrollmentRegistrationEntry.component.js
@@ -1,41 +1,14 @@
 // @flow
-import React, { useEffect } from 'react';
-import { useDispatch } from 'react-redux';
+import React from 'react';
 import { useScopeInfo } from '../../../hooks/useScopeInfo';
 import { scopeTypes } from '../../../metaData';
-import { startNewEnrollmentDataEntryInitialisation } from './EnrollmentRegistrationEntry.actions';
 import { EnrollmentDataEntry } from '../Enrollment';
 import { useCurrentOrgUnitInfo } from '../../../hooks/useCurrentOrgUnitInfo';
 import { useRegistrationFormInfoForSelectedScope } from '../common/useRegistrationFormInfoForSelectedScope';
 import type { OwnProps } from './EnrollmentRegistrationEntry.types';
 
-const useDataEntryLifecycle = (selectedScopeId, dataEntryId, scopeType) => {
-    const dispatch = useDispatch();
-    const { id: selectedOrgUnitId } = useCurrentOrgUnitInfo();
-    const { formId, formFoundation } = useRegistrationFormInfoForSelectedScope(selectedScopeId);
-    const registrationFormReady = !!formId;
-    useEffect(() => {
-        if (registrationFormReady && scopeType === scopeTypes.TRACKER_PROGRAM) {
-            dispatch(
-                startNewEnrollmentDataEntryInitialisation(
-                    { selectedOrgUnitId, selectedScopeId, dataEntryId, formFoundation },
-                ),
-            );
-        }
-    }, [
-        scopeType,
-        dataEntryId,
-        selectedScopeId,
-        selectedOrgUnitId,
-        registrationFormReady,
-        formFoundation,
-        dispatch,
-    ]);
-};
-
-export const EnrollmentRegistrationEntry = ({ selectedScopeId, id, ...rest }: OwnProps) => {
+export const EnrollmentRegistrationEntryComponent = ({ selectedScopeId, id, ...rest }: OwnProps) => {
     const { scopeType } = useScopeInfo(selectedScopeId);
-    useDataEntryLifecycle(selectedScopeId, id, scopeType);
     const { formId, registrationMetaData, formFoundation } = useRegistrationFormInfoForSelectedScope(selectedScopeId);
     const orgUnit = useCurrentOrgUnitInfo();
